Fix fetching EKS AMI release archive (#112)
This changelist fixes fetching the EKS AMI release archive by removing the check for adding a `v` prefix if it does not exist.

The last release published a [tag](https://github.com/awslabs/amazon-eks-ami/releases/tag/V20220123) that begins with an uppercase `V` and the code ends up prepending an additional `v`, rendering the URL incorrect. Rather than accounting for this case in code, this changelist removes that check altogether as `amazon-eks-ami` has been consistently adding the `v` prefix to their tags.